+++ src/OptiScheduledJob.ExtraParameters.Cms13/Attributes/ScheduledPlugInWithExtraParametersAttribute.cs
@@ -1,0 +1,10 @@
+using EPiServer.PlugIn;
+
+namespace OptiScheduledJob.ExtraParameters.Attributes
+{    
+    [AttributeUsage(AttributeTargets.Class)]
+    public class ScheduledPlugInWithExtraParametersAttribute : EPiServer.Scheduler.ScheduledJobAttribute
+    {
+        public Type ExtraParameterDefinition { get; set; }
+    }
+}