+++ src/OptiScheduledJob.ExtraParameters.Cms13/Models/ScheduledJobExtraParametersBase.cs
@@ -1,0 +1,7 @@
+namespace OptiScheduledJob.ExtraParameters.Models
+{
+    public abstract class ScheduledJobExtraParametersBase
+    {
+
+    }
+}