+++ src/OptiScheduledJob.ExtraParameters.Cms13/ViewModels/ExtraParametersViewModel.cs
@@ -1,0 +1,7 @@
+namespace OptiScheduledJob.ExtraParameters.ViewModels
+{
+    public class ExtraParametersViewModel
+    {
+        public IList<ExtraParameterSetting> ExtraParameterSettings { get; set; }
+    }
+}